fix: move multiprocessing_start_method_fork fixture to shared conftest
On Python 3.14, the default multiprocessing start method changed from
fork to forkserver (python/cpython#132898), which requires all Process
arguments to be picklable. The unprivileged_client Kubernetes client
object contains thread locks and closures that cannot be pickled,
causing _pickle.PicklingError in test_successful_concurrent_uploads.

Move the multiprocessing_start_method_fork fixture from
tests/chaos/conftest.py to tests/conftest.py so it can be reused
across test directories, and apply it to the affected upload test.

Changes:
- Move fixture to tests/conftest.py (shared across all test dirs)
- Remove duplicate from tests/chaos/conftest.py
- Add @pytest.mark.usefixtures("multiprocessing_start_method_fork")
  to test_successful_concurrent_uploads in test_upload.py
- Existing chaos consumers (test_standard, test_snapshot,
  test_migration) resolve the fixture from the parent conftest
